fix: Fix to calculate index for FFT bit reverse array

+++ src/SoundModule/Effectors/AudioWorkletProcessors/WebAssemblyModules/FFT.hpp
@@ -174,7 +174,7 @@ static void IFFT(float *const reals, float *const imags, const size_t size) {
     swap(reals, imags, index[k], k);
   }
 
-  for (int k = 0; k < size; k++) {
+  for (size_t k = 0; k < size; k++) {
     reals[k] /= size;
     imags[k] /= size;
   }
